conditionning drop

+++ lettings/migrations/0003_auto_20230528_1125.py
@@ -4,8 +4,11 @@ from django.db import migrations
 
 
 def delete_old_table(apps, schema_editor):
-    schema_editor.execute("DROP TABLE oc_lettings_site_address")
-    schema_editor.execute("DROP TABLE oc_lettings_site_letting")
+    table_names = schema_editor.connection.introspection.table_names()
+    if 'oc_lettings_site_address' in table_names:
+        schema_editor.execute("DROP TABLE oc_lettings_site_address")
+    if 'oc_lettings_site_letting' in table_names:
+        schema_editor.execute("DROP TABLE oc_lettings_site_letting")
 
 
 class Migration(migrations.Migration):
